add: Simple chat
Adds server and client.
Server handles connected people in a thread per client, stores clients in 'clients' array, which uses Mutex in order to avoid race conditions.

Example race condition that could happen is if we e.g. were announcing a message from a client and at the same time the client was removed.